Fix folly::coro::timeout to work with AsyncGenerator
Summary: The current signature results in an attempt to construct a `Try<semi_await_result_t<SemiAwaitable>>` from a `semi_await_try_result_t<SemiAwaitable>::element_type`. While the inner types match for Task, they don't for AsyncGenerator - the former is `AsyncGenerator<T>::NextResult`, which is not constructible from a T. (We could alternatively add that constructor, but that would result in timeout returning a `Task<AsyncGenerator<T>::NextResult>`, which is undesirably indirect).

Reviewed By: yfeldblum

Differential Revision: D28475605

fbshipit-source-id: c30a60e3ec942f801653a08b020796d47d4ef763