Add ConstructorCallback to AyncSocket
Summary:
Use the new ConstructorCallback class to allow exterior functions/classes
to get a callback for each new AsyncSocket()

Implications:
This adds an optional and an atomic int to the size of each AsyncSocket so
there is a very small increase in mem usage even if this is unused (which will
be most code, at least initially).

There is also a very small amount of additional CPU on constructing an
AsyncSocket: checking the std::option to see if it's non-nullopt.

Reviewed By: bschlinker

Differential Revision: D27516379

fbshipit-source-id: 56f4411afcd404dcf0092cce562230e5c0dcdf17