Add optional parameter to consume() and cancel()
Summary:
This diff adds an optional parameter to consume() and cancel(). This parameter will allow callback objects to know the source bridge that is calling them back. With this parameter, we will not need to create a separate callback object for each bridge (which saves memory).

The parameter is optional. If the parameter type is set to void, no parameter will be passed. This preserves backward compatibility with existing users of AtomicQueue that don't need a parameter.

Reviewed By: iahs

Differential Revision: D28550867

fbshipit-source-id: e7d998538c880c2a5c7649d3262cb7f8913e1439