Fix skip pointer population in EliasFanoCoding
Summary: Universe upper bound is inclusive, so we need to make sure that we have a pointer for `upperBitsUniverse / skipQuantum`. The bug has currently no effect as we use a tighter upper bound by reading the end of the list on construction, but it is necessary to fix this if we want to avoid doing that.

Reviewed By: yfeldblum, philippv, luciang

Differential Revision: D28886633

fbshipit-source-id: ba53fcc3f2ae1d52fe8e39576ad930754282ddd9