Change Context Pool Stripes from 4 to 128
Summary:
This diff increases the number of domains that locally cache a context from
4 to 128. In effect, this stops sharing contexts between cores, and instead
causes each core to have its own context. (If a machine has fewer than 128
CPUs, the unused slots will never be initialized.)

This diff is an experimental tuning. We don't expect it to have large positive
or negative effects, but we do currently see a non-zero amount of CPU spent
contending on the underlying pool's mutex. So this is an attempt to address
that. We'll look at CPU after it has rolled out and assess the results.

Reviewed By: terrelln

Differential Revision: D29944813

fbshipit-source-id: 7b10cb9dd2c49c02fad1f680a1bea02747990796