Add WorkersProvider interface to QueueObserver
Summary:
In order to colect backtraces from all threads consuming from a
task queue, we need the ability to expose their correpsonding thread IDs.
Here we introduce a new interface `WorkerProvider` which can be used to
collect the thread IDs. The implementations of this interface will come
later in the stack.

Reviewed By: mshneer

Differential Revision: D30101184

fbshipit-source-id: 24aaaeeba1983e3917c0224e94fc741b66a4f6a4