Record OS thread IDs for threads used in CPUThreadPoolExecutor
Summary:
In order to collect backtraces of all threads consuming from the executor's
queue(s), we need to record their actual thread IDs. We can keep track of the IDs
in a list owned by the executor. Each thread in the pool will add its own
thread ID to the tracking list by calling `folly::getOSThreadID()` and is also
responsible for removing it from the list as it exits.

Reviewed By: yfeldblum, mshneer

Differential Revision: D30103988

fbshipit-source-id: c59a1c31811684974fdb238433bd4796d0bbf2c8