add fdAttach callback
Summary: The fdAttach callback notifies all AsyncSocketLifecycleObservers attached to the socket that a file descriptor has just been associated with the socket. In AsyncSocket, the file descriptor is created when connect() is **called**, not when it succeeds. This callback lets observers know about the fd immediately; otherwise, observers would have to wait until the connect succeeds to learn about the file descriptor attached event.

Reviewed By: bschlinker

Differential Revision: D29910574

fbshipit-source-id: c452f1091c657a7d118b1e06546555aa04a34bf1