hazptr_obj_cohort: Fix reclamation of safe list
Summary:
Fix a bug in hazptr_obj_cohort handling of children of reclaimed nodes during reclamation of the safe list. When the cohort is active, the children nodes should be pushed to the domain (to be checked against hazard pointers) and not reclaimed immediately because it is possible that a child node is protected by hazard pointers.

Added test that catches the bug.

Reviewed By: yfeldblum

Differential Revision: D30170939

fbshipit-source-id: e5bfe46b6fa5d4e5ce5cef02a3934ced424138f0