Move fd_ to the initializer list in AsyncSocket constructor
Summary: When constructing a new AsyncSocket, we can use an already connected file descriptor, which we assign to the `fd_` member variable. We move this assignment from the constructor body to the constructor initializer list to avoid a situation where an observer is unaware of an attached fd because `fd_` was assigned after the observer was created.

Reviewed By: yfeldblum

Differential Revision: D30261699

fbshipit-source-id: 5adbc744715355fba08f5430a77c9094e246899d