Add a AsyncSocket::LifecycleObserver->connectError() callback
Summary:
The LifecycleObserver already tracks connect() but does not notify on
connectError().  Follow the same pattern.

Note that we intentionally added this to AsyncSocket's LifecycleObserver
rather than AsyncTransport b/c AsyncTransport is not supposed to have
a concept of a 'connection'.  This is counter to the 'connect()' callback
that already exists in AsyncTransport::LifecycleObserver so marked a quick
TODO to remember to move that.

Reviewed By: bschlinker

Differential Revision: D28612132

fbshipit-source-id: cefa650bd007d48e38f9cb67bc9a82d322036d4c