Fix Folly issues with -Wcovered-switch-default compiler flag
Summary:
When compiled with `-Wcovered-switch-default`, Folly fails to build because it adds `default` to switches which are already exhaustive. For C++ safety, these switches are intentionally including a default, so wrap them in a pragma block to indicate that and locally disable the warning.

Causes -fmodules import issues for upstream library in D30379442.

Still remaining issues with C++ definition: P445339268

Reviewed By: ispeters

Differential Revision: D30397871

fbshipit-source-id: adead9e01d0ba3b21c2e1446150cc6fed0f58602