folly: symbolizer: tests: switch from qsort to lfind -- ASAN in LLVM12 adds an __interceptor_qsort.part.0 which makes stacks differ between ASAN and non-ASAN
Summary:
LLVM 12's ASAN intercepts qsort and adds a new stack frame for the intercetor `__interceptor_qsort.part.0`.

That can be solved by switching offsets between mode/dev & mode/opt, but it only makes tests harder to read.

So switch to `lfind` which also lives in a separate `.so` (feature used in tests), but is simple enough that ASAN doesn't need to intercept it.

Differential Revision: D30665575

fbshipit-source-id: ba540aede33c37bea8e1f4777bd7325699f9a8d2