hazard pointers: Extend cleanup to cover cohort objects
Summary:
Extend cleanup to cover retired cohort objects.

The change includes incrementing num_bulk_reclaims_ before invoking do_reclamation either directly or in an executor. The corresponding decrement is done at the completion of do_reclamation, which may happen on a different thread if reclamation is done in an executor.

Reviewed By: davidtgoldblatt

Differential Revision: D30513628

fbshipit-source-id: 80227e85301274c66cc200ee5228654f1ca07d08