FanoutChannel: Change falcon code to use FanoutChannel's custom context
Summary: This diff changes the users of FanoutChannel to use the context functionality added in the previous diff. It includes a new helper class in falcon/proxy called ProxyNotificationFanoutChannel, which handles updating the current version and sending an initial NewSubscription to new subscribers with that version. This diff then changes ProxyCache, ProxyProdEntity, OverriddenBlobEntity, and OverriddenMapEntity to use ProxyNotificationFanoutChannel. (Previously, the logic in ProxyNotificationFanoutChannel was duplicated in all 4 of those classes.)

Reviewed By: aary

Differential Revision: D30889895

fbshipit-source-id: 39c9167f2bc6cc51e0b089eb5382faa498ffcb89