Make LogCategory::getPropagateLevelMessagesToParentRelaxed() const.
Summary:
There's no compelling reason I could uncover why someone with a const folly::LogCategory* shouldn't be able to get the propagation level, so I'm marking this method as const to allow for that.  This makes the method more consistent with all of the other get methods in the class.

For example, if one gets the category configuration map through getCategoryConfigs(), the map returned is const.

Reviewed By: yfeldblum

Differential Revision: D31066791

fbshipit-source-id: 46267ae7bb950fbd1b3c3d8b032ed2491d54e446