hazptr_test: Fix test and microbenchmark
Summary:
Fix reclamation_without_calling_cleanup test to wait for the start of asynchronous reclamation which is done on a different thread.

Fix list_protect_all_bench. The node links in the list must be immutable since the List destructor retires each node in the list, whereas nodes with mutable links automatically unlink their children and retire them if a child's last link is unlinked.

Reviewed By: yfeldblum

Differential Revision: D31101932

fbshipit-source-id: 3afa4b07cd56e07eccab9be69bfe895e4f54aee8