Rm non-const ctor for not_null
Summary:
D22214468 added a non-const copy constructor override. This does not compile on all platforms.

The reasoning in D22214468 pertains to overload resolution for copies. The specifics for adding the non-const copy constructor (in addition to having the const version) are lost. (If only we could ask the author... but past me does not have the answers that present me needs.) It is possible that later changes have obviated the need. Try removing.

Reviewed By: yfeldblum

Differential Revision: D31218444

fbshipit-source-id: 307030df044b4413125365782064066c99570f02