Promote DigestBuilder out of detail
Summary: `DigestBuilder` is a useful companion to `TDigest` (already public), has a simple and well-defined interface, better documentation than most classes in the parent directory, tests, and benchmarks. It doesn't need to be in `detail/`, and it can be useful for use cases not covered by `BufferedStat`, for example when sub-second resolution and a precise flush schedule are required.

Reviewed By: philippv, luciang

Differential Revision: D31643638

fbshipit-source-id: 4f2f54a933942cbe02d08cd63ab5fc04f13630e3